Compute bins

Compute bins values, middles, and data repartition over bins. Available options are:

"criteria"	(string)	Bins criteria: "equalwidth", "equalsize" or "leastsquare" (default).

"useFixedNb"	(bool)	TRUE to fix the number of bins, FALSE (default) to estimate it.

"fixedNb"	(int)	Fixed number of bins (default: 10).

"estimatedNb"	(pair<int,int>)	Bounds for bins number estimation (default: [5,30]).

"nbBinData"	(pair<int,int>)	Minimum and maximum number of data per bin for bins number estimation (default: [10,200]).

Usage

computeBins(data, options = list())

Arguments

data

(vector<double>) Input data.

options

(list) [optional] Computation options.

Value

A list bins values ("values"), middles ("middles") and the actual number of data per bin ("repartition").

Examples

if (FALSE) {
computeBins(data = c(1, 1.25, 2.5, 5, 5.5, 5.75, 7.5, 15, 16.5), options = list(nbBinData = c(1,10)))
}