[Monolix - PKanalix - Simulx] Compute the charts data

Compute (if needed) and export the charts data of a given plot or, if not specified, all the available project plots.

Usage

computeChartsData(plot = NULL, output = NULL, exportVPCSimulations = NULL)

Arguments

plot

(character) [optional][Monolix] Plot type. If not specified, all the available project plots will be considered. Available plots: bivariatedataviewer, covariateviewer, outputplot, indfits, obspred, residualsscatter, residualsdistribution, vpc, npc, predictiondistribution, parameterdistribution, randomeffects, covariancemodeldiagnosis, covariatemodeldiagnosis, likelihoodcontribution, fisher, saemresults, condmeanresults, likelihoodresults.

output

(character) [optional][Monolix] Plotted output (depending on the software, it can represent an observation, a simulation output, ...). By default, all available outputs are considered.

exportVPCSimulations

(bool) [optional][Monolix] Should VPC simulations be exported if available. Equals FALSE by default. NOTE: If 'plot" argument is not provided, 'output' and "task' arguments are ignored.

Details

computeChartsData can be used to compute and export the charts data for plots available in the graphical user interface as in Monolix, PKanalix or Simulx, when you export > export charts data.

The exported charts data is saved as txt files in the result folder, in the ChartsData subfolder.

Notice that it does not impact the current scenario.

To get a ggplot equivalent to the plot in the GUI, but customizable in R with the ggplot2 library, better use one of the plot... functions available in the connectors for Monolix and PKanalix (not available for Simulx). To get the charts data for one of these plot functions as a dataframe, you can use getChartsData.

See also

Examples

if (FALSE) {
computeChartsData() # Monolix - PKanalix - Simulx
computeChartsData(plot = "vpc", output = "y1") # Monolix
}