[Simulx] Get endpoint elements

Get the list of all endpoints.

Usage

getEndpoints()

Details

Endpoints summarize the outcome values over all individuals, for each simulation group and each replicate. Endpoints are defined and calculated as in Simulx GUI.

To compute the defined endpoints, use runEndpoints and get the results with getEndpointsResults.

To check if endpoints will be compared across simulation groups, use getGroupComparisonSettings. If group comparison is relevant, the way the comparison will be done for each endpoint (eg if statistical test and which p-value) should be defined in the endpoint element.

Each element is a list of

outcome (string or list) - Outcome on which the endpoint is based on. If one outcome, a string containing its name. If combined outcomes were used, a list with:

names (vector of strings) - Vector of outcome names

groupName (string) - Name of the outcome combination

operator (string) - Way in which output should be combined. One of "and"/"or" (in case of boolean outcomes) or "min"/"max" (in other cases)

metric (string) - Calculation method for the endpoint. One of "arithmeticMean", "geometricMean" or "median" if value-based outcome. In case of event-based outcomes, "kaplanMeier" (median survival) and in case of boolean outcomes, "percentTrue".

groupComparison (list) - Group comparison settings containing:

type (string) - One of "directComparison", "statisticalTest"

operator (string) - One of "!=", ">" or "<",

threshold (double) - A real number indicating the threshold for difference/oddsRatio

pvalue (double) - A real number indicating the p-value (if type is "statisticalTest")

See also

Examples