[Monolix] Get last estimated individual parameter values

Get the last estimated values for each subject of some of the individual parameters present within the current project.
WARNING: Estimated individual parameters values cannot be accessible until the individual estimation algorithm has been launched once.
NOTE: The user can choose to display only the individual parameter values estimated with a specific method.
Existing individual estimation methods :

Conditional Mean SAEM	"saem"

Conditional Mean	"conditionalMean"

Conditional Mode	"conditionalMode"

WARNING: Only the methods which have been used during the last scenario run can provide estimation results.

Usage

getEstimatedIndividualParameters(..., method = "")

Arguments

...

(string) Name of the individual parameters whose values must be displayed. Call getIndividualParameterModel to get a list of the individual parameters present within the current project.

method

[optional](string) Individual parameter estimation method whose results should be displayed. If there are latent covariate used in the model, the estimated modality is displayed too If this field is not specified, the results provided by all the methods used during the last scenario run are displayed.

Value

A data frame giving, for each wanted method, the last estimated values of the individual parameters of interest for each subject with the corresponding standard deviation values.

See also

Examples