[Simulx] Get remaining parameters for a simulation group

Get the values of the remaining parameters (typically the error model parameters) for a group.

Usage

getGroupRemaining(group)

Arguments

group

(character) Group name

Details

Remaining parameters are all parameters that appear in the structural model (in the input line of [LONGITUDINAL]) and are neither individual parameters not regressors. They are typically error model parameters.

If an individual parameters element is selected for simulation, and the model includes remaining parameters, it is possible to set their values with setGroupRemaining.

It typically enables to make a simulation with measurement noise, with an individual element. These error model parameters will impact the simulation only if a noisy observation (from the DEFINITION section of the [LONGITUDINAL] block) is set as output element (instead of a smooth prediction in OUTPUT or variable in EQUATION).

If a population parameters element is selected, it is not possible to set remaining parameters because these parameters are already part of the population element.

See also

Examples

  # get default values of error model parameters when importing a Monolix project
  initializeLixoftConnectors("monolix")
  monolix_project <- file.path(getDemoPath(), "1.creating_and_using_models", "1.1.libraries_of_models", "warfarinPK_project.mlxtran")
  initializeLixoftConnectors("simulx")
  importProject(monolix_project)
  setGroupElement(group = "simulationGroup1", elements = "mlx_EBEs")
  getGroupRemaining(group = "simulationGroup1")