[Monolix - PKanalix - Simulx] Get the name of a library model given a list of library filters.

Models can be loaded from a library based on a selection of filters as in PKanalix, Monolix and Simulx GUI. For a complete description of each model library, and guidelines on how to select models, please visit https://mlxtran.lixoft.com/model-libraries/.

Usage

getLibraryModelName(library, filters = list())

Arguments

library

(string) One of the MonolixSuite library of models. Possible values are "pk", "pd", "pkpd", "pkdoubleabs", "pm", "tmdd", "tte", "count" and "tgi".

filters

(list(name = string)) Named list of filters (optional), format: list(filterKey = "filterValue", ...). Default empty list. Since available filters are not in any particular order, filterKey should always be stated.

Value

Name of the filtered model, or vector of names of the available models if not all filters were selected. Names start with "lib:".

Details

getLibraryModelName enables to get the name of the model to be loaded. You can then use it in setStructuralModel or newProject to load the model in an existing or in a new project.

All possible keys and values for each of the libraries are listed below.

PK library

key		values

administration		bolus, infusion, oral, oralBolus

delay		noDelay, lagTime, transitCompartments

absorption		zeroOrder, firstOrder

distribution		1compartment, 2compartments, 3compartments

elimination		linear, MichaelisMenten

parametrization		rate, clearance, hybridConstants

bioavailability		true, false

PD library

key		values

response		immediate, turnover

drugAction		linear, logarithmic, quadratic, Emax, Imax, productionInhibition,

degradationInhibition, degradationStimulation, productionStimulation

baseline		const, 1-exp, exp, linear, null

inhibition		partialInhibition, fullInhibition

sigmoidicity		true, false

PKPD library

key		values

administration		bolus, infusion, oral, oralBolus

delay		noDelay, lagTime, transitCompartments

absorption		zeroOrder, firstOrder

distribution		1compartment, 2compartments, 3compartments

elimination		linear, MichaelisMenten

parametrization		rate, clearance

bioavailability		true, false

response		direct, effectCompartment, turnover

drugAction		Emax, Imax, productionInhibition, degradationInhibition,

degradationStimulation, productionStimulation

baseline		const, null

inhibition		partialInhibition, fullInhibition

sigmoidicity		true, false

PK double absorption library

key		values

firstAbsorption		zeroOrder, firstOrder

firstDelay		noDelay, lagTime, transitCompartments

secondAbsorption		zeroOrder, firstOrder

secondDelay		noDelay, lagTime, transitCompartments

absorptionOrder		simultaneous, sequential

forceLongerDelay		true, false

distribution		1compartment, 2compartments, 3compartments

elimination		linear, MichaelisMenten

parametrization		rate, clearance

Parent-metabolite library

key		values

administration		bolus, infusion, oral, oralBolus

firstPassEffect		noFirstPassEffect, withDoseApportionment,

withoutDoseApportionment

delay		noDelay, lagTime, transitCompartments

absorption		zeroOrder, firstOrder

transformation		unidirectional, bidirectional

parametrization		rate, clearance

parentDistribution		1compartment, 2compartments, 3compartments

parentElimination		linear, MichaelisMenten

metaboliteDistribution		1compartment, 2compartments, 3compartments

metaboliteElimination		linear, MichaelisMenten

TMDD library

key		values

administration		bolus, infusion, oral, oralBolus

delay		noDelay, lagTime, transitCompartments

absorption		zeroOrder, firstOrder

distribution		1compartment, 2compartments, 3compartments

tmddApproximation		MichaelisMenten, QE, QSS, full, Wagner,

constantRtot, constantRtotIB, irreversibleBinding

output		totalLigandLtot, freeLigandL

parametrization		rate, clearance

TTE library

key		values

tteModel		exponential, Weibull, Gompertz, loglogistic,

uniform, gamma, generalizedGamma

delay		true, false

numberOfEvents		singleEvent, repeatedEvents

typeOfEvent		intervalCensored, exact

dummyParameter		true, false

Count library

key		values

countDistribution		Poisson, binomial, negativeBinomial, betaBinomial,

generalizedPoisson, geometric, hypergeometric,

logarithmic, Bernoulli

zeroInflation		true, false

timeEvolution		constant, linear, exponential, Emax, Hill

parametrization		probabilityOfSuccess, averageNumberOfCounts

TGI library

key		values

shortcut		ClaretExponential, Simeoni, Stein, Wang,

Bonate, Ribba, twoPopulation

initialTumorSize		asParameter, asRegressor

kinetics		true, false

model		linear, quadratic, exponential, generalizedExponential,

exponentialLinear, Simeoni, Koch, logistic,

generalizedLogistic, SimeoniLogisticHybrid, Gompertz,

exponentialGompertz, vonBertalanffy, generalizedVonBertalanffy

additionalFeature		none, angiogenesis, immuneDynamics

treatment		none, pkModel, exposureAsRegressor, startAtZero,

startTimeAsRegressor, armAsRegressor

killingHypothesis		logKill, NortonSimon

dynamics		firstOrder, MichaelisMenten, MichaelisMentenHill,

exponentialKill, constant

resistance		ClaretExponential, resistantCells, none

delay		signalDistribution, cellDistribution, none

additionalTreatmentEffect		none, angiogenesisInhibition, immuneEffectorDecay

Examples

if (FALSE) {
getLibraryModelName(library = "pk", filters = list(administration = "oral", delay = "lagTime", absorption = "firstOrder", distribution = "1compartment", elimination = "linear", parametrization = "clearance"))
# returns "lib:oral1_1cpt_TlagkaVCl.txt"
getLibraryModelName("pd", list(response = "turnover", drugAction = "productionStimulation"))
# returns c("lib:turn_input_Emax.txt", "lib:turn_input_gammaEmax.txt")
}