[PKanalix] Get NCA parameter statistics

Get statistics over the estimated values of some of the NCA parameters of the current project. Statistics are computed on the different sets of individuals resulting from the stratification settings passed in argument or, if not given, the ones previously set.

Usage

getNCAParameterStatistics(parameters = c(), stratification = NULL)

Arguments

parameters

[optional](vector<string>) Name of the parameters whose values must be displayed and a list of information relative to these parameters (units & CDISC names). Possible parameters:

parameters related to the calculation of lambda_z: "Rsq", "Rsq_adjusted", "Corr_XY", "No_points_lambda_z", "Lambda_z", "Lambda_z_lower", "Lambda_z_upper", "HL_Lambda_z", "Lambda_z_intercept", "Span"

parameters calculated in case of plasma data: "Tlag", "T0", "Dose", "N_Samples", "C0", "Tmax", "Cmax", "Cmax_D", "Tlast", "Clast", "AUClast", "AUClast_D", "AUMClast", "MRTlast", "MRTlast", "AUCall", "AUCINF_obs", "AUCINF_D_obs", "AUC_PerCentExtrap_obs", "AUC_PerCentBack_Ext_obs", "AUMCINF_obs", "AUMC_PerCentExtrap_obs", "MRTINF_obs", "MRTINF_obs", "Vz_F_obs", "Cl_F_obs", "Vz_obs", "Cl_obs", "Vss_obs", "Clast_pred", "AUCINF_pred", "AUCINF_D_pred", "AUC_PerCentExtrap_pred", "AUC_PerCentBack_Ext_pred", "AUMCINF_pred", "AUMC_PerCentExtrap_pred", "MRTINF_pred", "MRTINF_pred", "Vz_F_pred", "Cl_F_pred", "Vz_pred", "Cl_pred", "Vss_pred"

parameters calculated in case of plasma data if partial AUC calculation intervals are provided through the partialAucTime: "AUC_lower_upper", "AUC_lower_upper_D", "CAVG_lower_upper"

parameters calculated only for multiple dose data: "Tau", "Ctau", "Ctrough", "AUC_TAU", "AUC_TAU_D", "AUC_TAU_PerCentExtrap", "AUMC_TAU", "Vz_F", "Vz", "CLss_F", "CLss", "Cavg", "FluctuationPerCent", "FluctuationPerCent_Tau", "Accumulation_Index", "Swing", "Swing_Tau", "Tmin", "Cmin", "Cmax", "MRTINF_obs"

parameters calculated in case of urine data: "T0", "Dose", "N_Samples", "Tlag", "Tmax_Rate", "Max_Rate", "Mid_Pt_last", "Rate_last", "AURC_last", "AURC_last_D", "Vol_UR", "Amount_Recovered", "Percent_Recovered", "AURC_all", "AURC_INF_obs", "AURC_PerCentExtrap_obs", "AURC_INF_pred", "AURC_PerCentExtrap_pred", "AURC_lower_upper", "Rate_last_pred"

parameters calculated in case of urine data if partial AUC calculation intervals are provided through the partialAucTime: "AURC_lower_upper", "AURC_lower_upper_D"

stratification

[optional] Stratification to apply on results. By default, project one is applied. Stratification is a list containing:

state Stratification state

groups Stratification groups list

See setNCAResultsStratification for more details about this argument structure.

See also

Examples