[Monolix - PKanalix - Simulx] Get current scenario

Get the list of tasks that will be run at the next call to runScenario. For Monolix, get in addition the associated method (linearization true or false), and the associated list of plots.

Usage

getScenario()

Value

The list of tasks that corresponds to the current scenario, indexed by task names.

Details

For Monolix, getScenario returns a given list of tasks, the linearization option and the list of plots.
Every task in the list is associated to a boolean.
NOTE: Within a MONOLIX scenario, the order according to which the different algorithms are run is fixed:

Algorithm	Algorithm Keyword

Population Parameter Estimation	"populationParameterEstimation"

Conditional Mode Estimation (EBEs)	"conditionalModeEstimation"

Sampling from the Conditional Distribution	"conditionalDistributionSampling"

Standard Error and Fisher Information Matrix Estimation	"standardErrorEstimation"

LogLikelihood Estimation	"logLikelihoodEstimation"

Plots	"plots"

For PKanalix, getScenario returns a given list of tasks. Every task in the list is associated to a boolean.
NOTE: Within a PKanalix scenario, the order according to which the different algorithms are run is fixed:

Algorithm	Algorithm keyword

Non Compartmental Analysis	"nca"

Bioequivalence estimation	"be"

For Simulx, setScenario returns a given list of tasks. Every task in the list is associated to a boolean.
NOTE: Within a Simulx scenario, the order according to which the different algorithms are run is fixed:

Algorithm	Algorithm keyword

Simulation	"simulation"

Outcomes and endpoints	"endpoints"

Note: every task can also be run separately with a specific function, such as runSimulation in Simulx, runEstimation in Monolix. The CA task in PKanalix cannot be part of a scenario, it must be run with runCAEstimation.

See also

Examples