[PKanalix] Generate NCA individual fits (elimination)

[PKanalix] Generate NCA individual fits (elimination)

Usage

plotNCAIndividualFits(
  data = NULL,
  settings = list(),
  preferences = list(),
  stratify = list()
)

Arguments

data

List of charts data as dataframe - Output of getChartsData ((getChartsData("plotNCAIndividualFits", ...)) If data not specified, charts data will be computed inside the function.

settings

List with the following settings

obsDots (bool) - If TRUE individual observations are displayed as dots (default TRUE).

obsLines (bool) - If TRUE individual observations are displayed as lines (default FALE).

cens (bool) - If TRUE censoring data are displayed as dots (default TRUE).

obsUnused (bool) - If TRUE and (and if dots is set to TRUE), individual observations not used for lambda z calculation are displayed as dots (default TRUE).

obsUnusedColor - If obsUnused is TRUE, unused data can be colored with the color used for observation (colored), or unused data can be colored in grey (greyed) (default greyed).

lambda_z (bool) - If TRUE individual fits are displayed (default TRUE).

dosingTimes (bool) - Add dosing times as vertical lines (default FALSE).

splitOccasions (bool) - If TRUE occasions are displayed on separate plots (default TRUE).

legend (bool) add (TRUE) / remove (FALSE) plot legend (default FALSE).

grid (bool) add (TRUE) / remove (FALSE) plot grid (default TRUE).

xlog (bool) add (TRUE) / remove (FALSE) log scaling on x axis (default FALSE).

ylog (bool) add (TRUE) / remove (FALSE) log scaling on y axis (default TRUE).

xlab (string) label on x axis (default "Time").

ylab (string) label on y axis (default "Concentration").

ncol (int) number of columns when facet = TRUE (default 4).

xlim (c(double, double)) limits of the x axis.

ylim (c(double, double)) limits of the y axis.

fontsize (integer) Plot text font size.

units (boolean) Set units in axis labels (default TRUE).

scales (string) Should scales be fixed ("fixed"), free ("free", the default), or free in one dimension ("free_x", "free_y") (default "free").

preferences

(optional) preferences for plot display, run getPlotPreferences("plotNCAIndividualFits") to check available displays.

stratify

List with the stratification arguments

ids - List of ids to display (by default all ids are displayed).

colorGroup - Color plots by groups of covariates (by default no color is applied). A list, or a list of list with fields:

name : The name of the covariate to use in grouping, or the name of the column id,

breaks : In case of a continuous covariate, a list of break values,

groups : [optional] In case of a categorical covariate, define groups of modalities.

filter - Filter data (by default no filtering is applied). A list, or a list of list with fields:

name - the name of the covariate to filter,

cat - in case of a categorical covariate, the name of the category to filter,

interval - in case of a continuous covariate, a list of filtering intervals.

colors - List of colors to use when colorGroup argument is defined

Value

A ggplot object

See also

Examples

  initializeLixoftConnectors(software = "pkanalix")
  project <- file.path(getDemoPath(), "2.case_studies/project_Theo_extravasc_SD.pkx")
  loadProject(project)

  runNCAEstimation()
  
  plotNCAIndividualFits()
 
  # display
  plotNCAIndividualFits(stratify = list(ids = c(1, 2, 3)),
                        settings = list(obsDots = T, lambda_z = T))
  plotNCAIndividualFits(stratify = list(ids = c(1, 2, 3)),
                        settings = list(obsDots = F, obsLines = T, lambda_z = T))
  
  # stratification
  plotNCAIndividualFits(stratify = list(ids = c(1, 2, 3, 4),
                                        filter = list(name = "Period", cat = 1)))
  plotNCAIndividualFits(stratify = list(ids = c(1, 2, 3, 4, 5),
                                        colorGroup = list(name = "SEQ"),
                                        colors = c("#5DC088", "#DBA92B")))
  plotNCAIndividualFits(stratify = list(colorGroup = list(list(name = "AGE", breaks = 25),
                                                          list(name = "Period"))))
  # update settings and preferences
  plotNCAIndividualFits(stratify = list(ids = c(1, 4)),
                        settings = list(ylog = TRUE, scales = "fixed"))
  plotNCAIndividualFits(settings = list(ncol = 5))
  plotNCAIndividualFits(settings = list(splitOccasions = FALSE, ncol = 5))
  preferences <- list(lambda_z = list(color = "pink", lineWidth = 1))
  plotNCAIndividualFits(stratify = list(ids = c(4, 5, 6)), preferences = preferences)
  
  # pre compute dataset
  data <- getChartsData(plot = "plotNCAIndividualFits", ids = c(1, 2))
  plotNCAIndividualFits(data = data)