[Monolix] Plot distribution of the predictions

Note that computation settings are not available for this connector in 2021 version: Number of bands is set to 9 and Level is set to 90

Usage

plotPredictionDistribution(
  obsName = NULL,
  settings = list(),
  preferences = list(),
  data = NULL
)

Arguments

obsName

(string) Name of the observation (in dataset header). By default the first observation is considered.

settings

a list of optional settings

perc (bool) - If TRUE display 9 Bands for each percentile (default TRUE).

median (bool) - If TRUE display Median (default TRUE).

obs (bool) - If TRUE display observations as dots (default FALSE).

cens (bool) - If TRUE display censored observations as dots (default FALSE).

binLimits (bool) If TRUE display limits of bins (default FALSE). For discrete data only.

legend (bool) add (TRUE) / remove (FALSE) plot legend (default FALSE).

grid (bool) add (TRUE) / remove (FALSE) plot grid (default TRUE).

xlog (bool) add (TRUE) / remove (FALSE) log scaling on x axis (default FALSE).

ylog (bool) add (TRUE) / remove (FALSE) log scaling on y axis (default FALSE).

xlab (string) label on x axis (default "Time").

ylab (string) label on y axis (default obsName).

ncol (int) number of columns when facet = TRUE (default 4).

xlim (c(double, double)) limits of the x axis.

ylim (c(double, double)) limits of the y axis.

fontsize (integer) Plot text font size.

scales (string) Should scales be fixed ("fixed"), free ("free", the default), or free in one dimension ("free_x", "free_y") (default "free").

preferences

(optional) preferences for plot display, run getPlotPreferences("plotPredictionDistribution") to check available displays.

data

List of charts data as dataframe - Output of getChartsData (getChartsData("plotPredictionDistribution", ...)) If data not specified, charts data will be computed inside the function.

Value

a ggplot2 object

Details

Note that stratification options are not available for this connector in 2021 version:

See also

Examples

  initializeLixoftConnectors(software = "monolix")
  
  # continuous data
  project <- file.path(getDemoPath(), "1.creating_and_using_models",
                       "1.1.libraries_of_models", "theophylline_project.mlxtran")
  loadProject(project)

  runPopulationParameterEstimation()

  plotPredictionDistribution()