[Monolix] Generate Scatter plots of the residuals

Note that 'prediction interval' setting is not available in 2021 version for this connector.

Usage

plotResidualsScatterPlot(
  obsName = NULL,
  residuals = c("indiv"),
  xaxis = c("time", "prediction"),
  settings = list(),
  preferences = list(),
  stratify = list(),
  data = NULL
)

Arguments

obsName

(string) Name of the observation (in dataset header). By default the first observation is considered.

residuals

(string) List of residuals to display: population residuals ("pop"), individual residuals ("indiv"), normalized prediction distribution error ("npde") (default c("indiv")).

xaxis

(string) List of x-axis to display: time ("time"), prediction ("prediction") (default c("time", "prediction") for continuous data, c("time") for discrete data).

settings

List with the following settings

indivEstimate (string) Calculation of individual estimates: conditional mean ("mean"), conditional mode with EBE's ("mode"), conditional distribution ("simulated") (default "mode"). For continuous data only

level (int) level for prediction intervals computation (default 90).

higherPercentile (int) Higher percentile for empirical and predicted percentiles computation (default 90).

useCensored (bool) Choose to use BLQ data (TRUE) or to ignore it (FALSE) to compute the statistics (default TRUE). For continuous data only.

censoring (string) BLQ data can be simulated ('simulated'), or can be equal to the limit of quantification ('loq') (default 'simulated'). For continuous data only.

binsSettings a list of settings for bins:

criteria (string) Bining criteria, one of 'equalwidth', 'equalsize', or 'leastsquare' methods. (default leastsquare).

is.fixedNbBins (bool) If TRUE define a fixed number of bins, else define a range for automatic selection (default TRUE).

nbBins (int) Define a fixed number of bins (default 10).

binRange (vector(int, int)) Define a range for the number of bins (default c(5, 100)).

nbBinData (vector(int, int)) Define a range for the number of data points per bin (default c(10, 200)).

residuals (bool) - If TRUE display residuals (default TRUE).

cens (bool) - If TRUE display censored data (default TUE).

empPercentiles (bool) - If TRUE display empirical percentiles (default FALSE).

predPercentiles (bool) - If TRUE display predicted percentiles (default FALSE).

spline (bool) - If TRUE display spline (default FALSE).

binLimits (bool) - If TRUE Add bins limits as vertical lines (default FALSE).

legend (bool) add (TRUE) / remove (FALSE) plot legend (default FALSE).

grid (bool) add (TRUE) / remove (FALSE) plot grid (default TRUE).

xlog (bool) add (TRUE) / remove (FALSE) log scaling on x axis (default FALSE).

ylog (bool) add (TRUE) / remove (FALSE) log scaling on y axis (default FALSE).

ncol (int) number of columns when facet = TRUE (default 4).

xlim (c(double, double)) limits of the x axis.

ylim (c(double, double)) limits of the y axis.

fontsize (integer) Plot text font size.

scales (string) Should scales be fixed ("fixed"), free ("free", the default), or free in one dimension ("free_x", "free_y") (default "free").

preferences

(optional) preferences for plot display, run getPlotPreferences("plotResidualsScatterPlot") to check available displays.

stratify

List with the stratification arguments

ids - List of ids to display (by default all ids are displayed).

splitGroup - Split plots by groups of covariates (by default no split is applied). A list, or a list of list with fields:

name : The name of the covariate to use in grouping,

breaks : In case of a continuous covariate, a list of break values,

groups : [optional] In case of a categorical covariate, define groups of modalities.

colorGroup - Color plots by groups of covariates (by default no color is applied). A list, or a list of list with fields:

name : The name of the covariate to use in grouping, or the name of the column id,

breaks : In case of a continuous covariate, a list of break values,

groups : [optional] In case of a categorical covariate, define groups of modalities.

filter - Filter data (by default no filtering is applied). A list, or a list of list with fields:

name - the name of the covariate to filter,

cat - in case of a categorical covariate, the name of the category to filter,

interval - in case of a continuous covariate, a list of filtering intervals.

colors - List of colors to use when colorGroup argument is defined

data

List of cahrts data as dataframe - Output of getChartsData (getChartsData("plotResidualsScatterPlot", ...)) If data not specified, charts data will be computed inside the function.

Value

A ggplot object if one prediction type,

A TableGrob object if multiple plots (output of grid.arrange)

See also

Examples

  initializeLixoftConnectors(software="monolix")
  project <- file.path(getDemoPath(), "1.creating_and_using_models",
                       "1.1.libraries_of_models", "theophylline_project.mlxtran")
  loadProject(project)

  runPopulationParameterEstimation()
  runConditionalDistributionSampling()
  runConditionalModeEstimation()
  
  plotResidualsScatterPlot()
  plotResidualsScatterPlot(residuals="indiv", settings=list(indivEstimate="simulated"))
  plotResidualsScatterPlot(residuals="indiv", settings=list(indivEstimate="mode"))
  plotResidualsScatterPlot(xaxis="prediction", residuals="pop")
  plotResidualsScatterPlot(xaxis="time", residuals="pop")
  plotResidualsScatterPlot(residuals="npde")
  plotResidualsScatterPlot(settings=list(spline=T))
  plotResidualsScatterPlot(settings=list(empPercentiles=T, level=90,
                                         binsSettings=list(is.fixedNbBins=T, nbBins=5),
                                         binLimits=T))
                          
  # Stratification
  plotResidualsScatterPlot(stratify=list(filter=list(name="SEX", cat="F")))
  plotResidualsScatterPlot(stratify=list(splitGroup=list(name="WEIGHT", breaks=c(75))))
  plotResidualsScatterPlot(stratify=list(colorGroup=list(name="WEIGHT", breaks=c(75))))
  
  data <- getChartsData(plotName="plotResidualsScatterPlot",
                        computeSettings=list(indivEstimate="simulated"))
  plotResidualsScatterPlot(data=data)
  
  plotResidualsScatterPlot(residuals=c("indiv", "pop"),
                           settings=list(indivEstimate="simulated"))
  plotResidualsScatterPlot(residuals="indiv", xaxis=c("prediction"),
                           settings=list(indivEstimate="mode"))
  plotResidualsScatterPlot(xaxis=c("prediction"), residuals=c("indiv", "pop"))
  plotResidualsScatterPlot(residuals="npde")