[PKanalix] Set NCA results stratification

Set the stratification used to compute NCA parameters stratistics table. Stratification is defined by:

stratification covariate groups which are shared by both NCA and CA results

a stratification state which is specific to each task results

Usage

setNCAResultsStratification(
  split = NULL,
  filter = NULL,
  groups = NULL,
  state = NULL
)

Arguments

split

(vector<string>) Ordered list of splitted covariates

filter

(list< pair<string, vector<int>> >) List of paired containing a covariate name and the indexes of associated kept groups

groups

Stratification groups list

state

Stratification state

Details

For each covariate, stratification groups can be defined as a list with:

name	string	covariate name

definition	vector<double>(continuous) || list<vector<string>>(categorical)	group separations (continuous) || modality sets (categorical)

A stratification state is represented as a list with:

split	vector<string>	ordered list of splitted covariates

filter	list< pair<string, vector<int>> >	list of paired containing a covariate name and the indexes of associated kept groups

Note: For acceptance criteria filtering, it is possible to give only the criterion name instead of a pair.

See also

Examples

if (FALSE) {
setNCAResultsStratification(split = "SEX")
setNCAResultsStratification(split = c("SEX", "WEIGHT"))

setNCAResultsStratification(filter = "Span")
setNCAResultsStratification(filter = list("Span", list("SEX", 1)))

setNCAResultsStratification(split = "WEIGHT", filter = list(list("TRT", c(1,2))),
groups = list(list(name = "WEIGHT", definition = c(65,5, 72)), list(name = "TRT", definition = list(c("a","b"), "c", c("d","e")))))

s = getNCAResultsStratification()
setNCAResultsStratification(state = s$state, groups = s$groups)
}